🏗️ Architect: Modularized FlipMirror equipment
Refactored the `apts/opticalequipment/flip_mirror.py` file into a modular package structure.
- Moved core `FlipMirror` logic to `base.py`.
- Distributed 21 database entries and factory methods across 18 vendor-specific modules in a `vendors/` subdirectory.
- Implemented dynamic aggregation in `__init__.py` to maintain the public API and backward compatibility.
- Updated relative imports to account for the new package nesting.
- Verified with full unit test suite (663 tests passed).
