+++ apts/opticalequipment/flip_mirror/vendors/vixen.py
@@ -1,0 +1,23 @@
+from ..base import FlipMirror
+
+
+class VixenFlipMirror(FlipMirror):
+    _DATABASE = {
+        "Vixen_Flipmirror": {
+            "brand": "Vixen",
+            "name": "Flipmirror",
+            "type": "type_flip_mirror",
+            "optical_length": 0,
+            "mass": 460,
+            "tside_thread": '2"',
+            "tside_gender": "Male",
+            "cside_thread": '1.25"',
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def Vixen_Flipmirror(cls):
+        return cls.from_database(cls._DATABASE["Vixen_Flipmirror"])
